[TAN-3815] Created native_survey_method classes
