+++ back/lib/native_survey_method/community_monitor.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module NativeSurveyMethod
+  class CommunityMonitor < Base
+    # TODO: implement different methods for community monitor
+  end
+end
