+++ back/lib/participation_method/native_survey.rb
@@ -2,14 +2,10 @@
 
 module ParticipationMethod
   class NativeSurvey < Base
+    delegate :allowed_extra_field_input_types, :default_fields, to: :native_survey_method
+
     def self.method_str
       'native_survey'
-    end
-
-    def allowed_extra_field_input_types
-      %w[page number linear_scale rating text multiline_text select multiselect
-        multiselect_image file_upload shapefile_upload point line polygon
-        ranking matrix_linear_scale]
     end
 
     def assign_defaults(input)
@@ -38,51 +34,6 @@ module ParticipationMethod
 
     def custom_form
       phase.custom_form || CustomForm.new(participation_context: phase)
-    end
-
-    def default_fields(custom_form)
-      return [] if custom_form.persisted?
-
-      multiloc_service = MultilocService.new
-      [
-        CustomField.new(
-          id: SecureRandom.uuid,
-          key: 'page1',
-          resource: custom_form,
-          input_type: 'page',
-          page_layout: 'default'
-        ),
-        CustomField.new(
-          id: SecureRandom.uuid,
-          key: CustomFieldService.new.generate_key(
-            multiloc_service.i18n_to_multiloc('form_builder.default_select_field.title').values.first
-          ),
-          resource: custom_form,
-          input_type: 'select',
-          title_multiloc: multiloc_service.i18n_to_multiloc('form_builder.default_select_field.title'),
-          options: [
-            CustomFieldOption.new(
-              id: SecureRandom.uuid,
-              key: 'option1',
-              title_multiloc: multiloc_service.i18n_to_multiloc('form_builder.default_select_field.option1')
-            ),
-            CustomFieldOption.new(
-              id: SecureRandom.uuid,
-              key: 'option2',
-              title_multiloc: multiloc_service.i18n_to_multiloc('form_builder.default_select_field.option2')
-            )
-          ]
-        ),
-        CustomField.new(
-          id: SecureRandom.uuid,
-          key: 'survey_end',
-          resource: custom_form,
-          input_type: 'page',
-          page_layout: 'default',
-          title_multiloc: multiloc_service.i18n_to_multiloc('form_builder.form_end_page.title_text_3'),
-          description_multiloc: multiloc_service.i18n_to_multiloc('form_builder.form_end_page.description_text_3')
-        )
-      ]
     end
 
     # Survey responses do not have a fixed field that can be used
@@ -135,6 +86,10 @@ module ParticipationMethod
 
     def supports_toxicity_detection?
       false
+    end
+
+    def native_survey_method
+      Factory.instance.native_survey_method_for(phase)
     end
   end
 end
